% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena128.R
\docType{data}
\name{lena128}
\alias{lena128}
\title{lena image at size of \eqn{(128 \times 128)}}
\format{
matrix of size \eqn{(128\times 128)}
}
\source{
\href{http://sipi.usc.edu/database/?volume=misc}{USC SIPI Image Database}
}
\usage{
data(lena128)
}
\description{
\emph{Lena} is probably one of the most well-known example in image processing and computer vision.
Well, here is a brief introduction on \href{http://www.cs.cmu.edu/~chuck/lennapg/lenna.shtml}{the story of Lena}.
}
\examples{
data(lena128)
image(lena128, col=gray((0:100)/100), axes=FALSE, main="lena128")
}
\references{
Gonzalez, Rafael C. and Woods, Richard E. (2017) \emph{Digital Image Processing} (4th ed.). ISBN 0133356728.
}
\keyword{datasets}
