% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lena256.R
\docType{data}
\name{lena256}
\alias{lena256}
\title{lena image at size of \eqn{(256 \times 256)}}
\format{
matrix of size \eqn{(256\times 256)}
}
\source{
\href{http://sipi.usc.edu/database/?volume=misc}{USC SIPI Image Database}
}
\usage{
data(lena256)
}
\description{
\emph{Lena} is probably one of the most well-known example in image processing and computer vision.
Well, here is a brief introduction on \href{http://www.cs.cmu.edu/~chuck/lennapg/lenna.shtml}{the story of Lena}.
}
\examples{
data(lena256)
image(lena256, col=gray((0:100)/100), axes=FALSE, main="lena256")
}
\references{
Gonzalez, Rafael C. and Woods, Richard E. (2017) \emph{Digital Image Processing} (4th ed.). ISBN 0133356728.
}
\keyword{datasets}
